Added encoding override when downloading articles.

+++ scripts/download_articles.py
@@ -1,9 +1,13 @@
 import argparse
+import json
 import os
+import traceback
 from datetime import datetime
 
 import psycopg2
 import requests
+
+from scripts.gather_articles_metadata import GatherArticlesMetadata
 
 
 class DownloadArticles:
@@ -16,9 +20,14 @@ class DownloadArticles:
     @staticmethod
     def parse_commandline():
         parser = argparse.ArgumentParser(formatter_class=argparse.ArgumentDefaultsHelpFormatter)
+        parser.add_argument('--dry-run', action='store_true', default=False,
+                            help='Don\'t store output and print it to stdout.')
+        parser.add_argument('--id', type=int, default=None, help='Specify id of article to download.')
         return parser.parse_args()
 
     def run(self):
+        domain_types = GatherArticlesMetadata._init_domain_types()
+
         cur = self.db_con.cursor()
 
         current_date = datetime.now()
@@ -31,14 +40,17 @@ class DownloadArticles:
             current_date.minute,
             current_date.second)
 
-        if not os.path.isdir(self.FOLDER_PREFIX):
-            os.makedirs(self.FOLDER_PREFIX)
-        if not os.path.isdir(folder_name):
-            os.makedirs(folder_name)
+        if not self.args.dry_run:
+            if not os.path.isdir(folder_name):
+                os.makedirs(folder_name)
 
-        cur.execute('SELECT a.id, a.website_domain, a.url FROM article_metadata a '
-                    'LEFT OUTER JOIN article_raw_html r ON r.article_metadata_id = a.id '
-                    'WHERE r.article_metadata_id IS NULL')
+        if self.args.id:
+            cur.execute('SELECT a.id, a.website_domain, a.url FROM article_metadata a '
+                        'WHERE a.id = %s', (self.args.id, ))
+        else:
+            cur.execute('SELECT a.id, a.website_domain, a.url FROM article_metadata a '
+                        'LEFT OUTER JOIN article_raw_html r ON r.article_metadata_id = a.id '
+                        'WHERE r.article_metadata_id IS NULL')
         articles_metadata = cur.fetchall()
         for index, (id, website_domain, url) in enumerate(articles_metadata):
             filename = '%s_%s.html' % (id, website_domain.replace('/', '-'))
@@ -48,16 +60,26 @@ class DownloadArticles:
                 response = requests.get(url, timeout=15, headers={
                     'User-Agent': 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_14_2) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/71.0.3578.98 Safari/537.36'
                 })
+                for domain_type in domain_types:
+                    if domain_type.get_name() == website_domain:
+                        response.encoding = domain_type.get_encoding()
+                        break
 
-                with open(full_path, 'w') as file:
-                    file.write(response.text)
+                if self.args.dry_run:
+                    print(response.text)
+                    print(response.encoding)
+                    print(full_path)
+                else:
+                    with open(full_path, 'w') as file:
+                        file.write(response.text)
 
-                cur.execute(
-                    'INSERT INTO article_raw_html (article_metadata_id, filename) VALUES (%s, %s)',
-                    (id, full_path))
-                self.db_con.commit()
+                    cur.execute(
+                        'INSERT INTO article_raw_html (article_metadata_id, filename) VALUES (%s, %s)',
+                        (id, full_path))
+                    self.db_con.commit()
                 print('(%s/%s) Finished downloading: %s.' % (index + 1, len(articles_metadata), url))
             except Exception as e:
+                traceback.print_exc()
                 print('(%s/%s) Error downloading: %s with message %s.' % (index + 1, len(articles_metadata), url, e))
 
         cur.close()
